# Change filter_queryset_by_group to do filters in one operation