--- a/inventory/views.py
+++ b/inventory/views.py
@@ -127,11 +127,6 @@ class GroupMixin():
         if self.group_instance:
             filter_path = self.access_filter[queryset.model][self.classes[group_type]]
 
-            # If filter_path is Machine there is nothing to filter on.
-            if filter_path:
-                kwargs = {filter_path: self.group_instance}
-                queryset = queryset.filter(**kwargs)
-
         # Remove undeployed machines from the results.
         # It's important that we are filtering for deployed machines
         # here rather than excluding undeployed machines-you get
@@ -143,7 +138,12 @@ class GroupMixin():
         # construct the keyword argument name to filter.
         deployed_filter = '{}{}deployed'.format(
             self.access_filter[queryset.model][Machine], '' if queryset.model is Machine else '__')
-        queryset = queryset.filter(**{deployed_filter: True})
+        # If filter_path is Machine there is nothing to filter on.
+        if filter_path:
+            kwargs = {filter_path: self.group_instance}
+            queryset = queryset.filter(**kwargs, **{deployed_filter: True})
+        else:
+            queryset = queryset.filter(**{deployed_filter: True})
 
         return queryset